move autocompletion helper methods to specific strategies
these were only used in one strategy, so no need to share via inheritance
