+++ src/services/ghost/strategies/NewLineCompletionStrategy.ts
@@ -270,4 +270,71 @@ Important:
 		}
 		return false
 	}
+
+	/**
+	 * Formats recent operations for inclusion in prompts
+	 */
+	private formatRecentOperations(operations: any[]): string {
+		if (!operations || operations.length === 0) return ""
+
+		let result = "## Recent Actions\n"
+		operations.slice(0, 5).forEach((op, index) => {
+			result += `${index + 1}. ${op.description}\n`
+			if (op.content) {
+				result += `   \`\`\`\n   ${op.content}\n   \`\`\`\n`
+			}
+		})
+
+		return result
+	}
+
+	/**
+	 * Helper to check if a line appears to be incomplete
+	 */
+	private isIncompleteStatement(line: string): boolean {
+		const trimmed = line.trim()
+
+		// Check for common incomplete patterns
+		const incompletePatterns = [
+			/^(if|else if|while|for|switch|try|catch)\s*\(.*\)\s*$/, // Control structures without body
+			/^(function|class|interface|type|enum)\s+\w+.*[^{]$/, // Declarations without body
+			/[,\+\-\*\/\=\|\&]\s*$/, // Operators at end
+			/^(const|let|var)\s+\w+\s*=\s*$/, // Variable declaration without value
+			/\.\s*$/, // Property access incomplete
+			/\(\s*$/, // Opening parenthesis
+			/^\s*\.\w*$/, // Method chaining incomplete
+		]
+
+		return incompletePatterns.some((pattern) => pattern.test(trimmed))
+	}
+
+	/**
+	 * Gets surrounding code context (lines before and after cursor)
+	 */
+	private getSurroundingCode(
+		document: TextDocument,
+		range: Range,
+		linesBefore: number = 10,
+		linesAfter: number = 10,
+	): { before: string; after: string; currentLine: string } {
+		const currentLineNum = range.start.line
+		const startLine = Math.max(0, currentLineNum - linesBefore)
+		const endLine = Math.min(document.lineCount - 1, currentLineNum + linesAfter)
+
+		let before = ""
+		let after = ""
+		const currentLine = document.lineAt(currentLineNum).text
+
+		// Get lines before cursor
+		for (let i = startLine; i < currentLineNum; i++) {
+			before += document.lineAt(i).text + "\n"
+		}
+
+		// Get lines after cursor
+		for (let i = currentLineNum + 1; i <= endLine; i++) {
+			after += document.lineAt(i).text + "\n"
+		}
+
+		return { before, after, currentLine }
+	}
 }
